+++ CHANGELOG.md
@@ -36,6 +36,7 @@
   Rackspace plugin rescues Errno::ENOENT if xenstor-* utils are not found (OHAI-587)
   
 
+* root_group provider not implemented for Windows (OHAI-491)
 * `Ohai::Exceptions::AttributeNotFound` errors in Chef's ohai resource
 * Be reluctant to call something an LXC host (OHAI-573)